Change image patch generation to disclude mean 0 images; InBreast 128x128, here wego

+++ data/SAGE.py
@@ -24,7 +24,7 @@ def generate_breast_patch(flags, image_dict):
                 x = np.random.randint(low=0, high=dims[0] - flags['image_dim'])
                 y = np.random.randint(low=0, high=dims[1] - flags['image_dim'])
                 img = image[x:x + flags['image_dim'], y:y + flags['image_dim']]
-                if img.mean < 0.001:
+                if float(img.mean) < float(0.001):
                     b -= 1
                     continue
                 img = img / img.max()
